Move home page to component
