+++ src/components/index.ts
@@ -3,6 +3,7 @@ export { FeatureBlock } from "./FeatureBlock/FeatureBlock";
 export { Footer } from "./Footer/Footer";
 export { Header } from "./Header/Header";
 export { HeroSection } from "./HeroSection/HeroSection";
+export { HomePage } from "./HomePage/HomePage";
 export { ImageTextBlock } from "./ImageTextBlock/ImageTextBlock";
 export { Layout } from "./Layout/Layout";
 export { SEO } from "./SEO/SEO";
